requirements.txt
google-adk==1.0.0
llama-index==0.12.35
